Move SharableService to services package
