+++ lib/galaxy/webapps/galaxy/services/histories.py
@@ -32,7 +32,6 @@ from galaxy.managers.histories import (
     HistoryManager,
     HistorySerializer,
 )
-from galaxy.managers.sharable import ShareableService
 from galaxy.managers.users import UserManager
 from galaxy.schema import (
     FilterQueryParams,
@@ -55,6 +54,7 @@ from galaxy.schema.types import LatestLiteral
 from galaxy.security.idencoding import IdEncodingHelper
 from galaxy.util import restore_text
 from galaxy.webapps.galaxy.services.base import ServiceBase
+from galaxy.webapps.galaxy.services.sharable import ShareableService
 
 log = logging.getLogger(__name__)
